+++ src/main/java/me/jinjjahalgae/domain/user/usecase/delete/myaccount/DeleteMyAccountUseCase.java
@@ -1,4 +1,4 @@
-package me.jinjjahalgae.domain.user.usecase.delete_my_account;
+package me.jinjjahalgae.domain.user.usecase.delete.myaccount;
 
 public interface DeleteMyAccountUseCase {
     void execute(Long userId);
