+++ src/main/java/me/jinjjahalgae/domain/user/usecase/update/myinfo/UpdateMyInfoUseCase.java
@@ -1,7 +1,7 @@
-package me.jinjjahalgae.domain.user.usecase.update_my_info;
+package me.jinjjahalgae.domain.user.usecase.update.myinfo;
 
 import me.jinjjahalgae.domain.user.usecase.common.dto.MyInfoResponse;
-import me.jinjjahalgae.domain.user.usecase.update_my_info.dto.UpdateMyInfoRequest;
+import me.jinjjahalgae.domain.user.usecase.update.myinfo.dto.UpdateMyInfoRequest;
 
 public interface UpdateMyInfoUseCase {
     MyInfoResponse execute(Long userId, UpdateMyInfoRequest request);
