+++ src/main/java/me/jinjjahalgae/domain/user/usecase/update/myinfo/UpdateMyInfoUseCaseImpl.java
@@ -1,11 +1,11 @@
-package me.jinjjahalgae.domain.user.usecase.update_my_info;
+package me.jinjjahalgae.domain.user.usecase.update.myinfo;
 
 import lombok.RequiredArgsConstructor;
 import me.jinjjahalgae.domain.user.User;
 import me.jinjjahalgae.domain.user.UserRepository;
 import me.jinjjahalgae.domain.user.mapper.UserMapper;
 import me.jinjjahalgae.domain.user.usecase.common.dto.MyInfoResponse;
-import me.jinjjahalgae.domain.user.usecase.update_my_info.dto.UpdateMyInfoRequest;
+import me.jinjjahalgae.domain.user.usecase.update.myinfo.dto.UpdateMyInfoRequest;
 import me.jinjjahalgae.global.exception.ErrorCode;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
