+++ src/main/java/me/jinjjahalgae/presentation/api/AuthController.java
@@ -5,9 +5,9 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import me.jinjjahalgae.domain.auth.usecase.social_login.dto.SocialLoginRequest;
-import me.jinjjahalgae.domain.auth.usecase.social_login.dto.SocialLoginResponse;
-import me.jinjjahalgae.domain.auth.usecase.social_login.SocialLoginUseCase;
+import me.jinjjahalgae.domain.auth.usecase.login.social.dto.SocialLoginRequest;
+import me.jinjjahalgae.domain.auth.usecase.login.social.dto.SocialLoginResponse;
+import me.jinjjahalgae.domain.auth.usecase.login.social.SocialLoginUseCase;
 import me.jinjjahalgae.domain.auth.usecase.logout.LogoutUseCase;
 import me.jinjjahalgae.domain.auth.usecase.refresh.RefreshUseCase;
 import me.jinjjahalgae.domain.auth.usecase.refresh.dto.RefreshRequest;
