+++ src/main/java/me/jinjjahalgae/presentation/api/docs/auth/AuthControllerDocs.java
@@ -10,8 +10,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletResponse;
-import me.jinjjahalgae.domain.auth.usecase.social_login.dto.SocialLoginRequest;
-import me.jinjjahalgae.domain.auth.usecase.social_login.dto.SocialLoginResponse;
+import me.jinjjahalgae.domain.auth.usecase.login.social.dto.SocialLoginRequest;
+import me.jinjjahalgae.domain.auth.usecase.login.social.dto.SocialLoginResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 import me.jinjjahalgae.global.exception.ErrorResponse;
 import me.jinjjahalgae.global.security.jwt.CustomJwtPrincipal;
