+++ src/main/java/me/jinjjahalgae/presentation/api/docs/auth/login/SocialLoginBodySwaggerResponse.java
@@ -1,6 +1,6 @@
 package me.jinjjahalgae.presentation.api.docs.auth.login;
 
-import me.jinjjahalgae.domain.auth.usecase.social_login.dto.SocialLoginResponse;
+import me.jinjjahalgae.domain.auth.usecase.login.social.dto.SocialLoginResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 
 public class SocialLoginBodySwaggerResponse extends CommonResponse<SocialLoginResponse> {
